moved reusable function and classes from 2020-09-18 to processing.py
